pyproject.toml
[tool.poetry]
name = "cheat-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.2.0"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
mysql-connector-python = "^8.2.0"
aiomysql = "^0.2.0"
mysql = "^0.0.3"
alembic = "^1.13.0"
boto3 = "^1.34.4"
requests = "^2.31.0"
redis = "^5.0.1"
databases = "^0.8.0"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
